requirements.txt
mysql-connector-python==8.0.19
requests==2.23.0
